+++ BE/amr_control_system/src/main/java/com/sfaas/amr_control_system/repository/WipLotRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WipLotRepository extends JpaRepository<WipLot, Integer> {
+public interface WipLotRepository extends JpaRepository<WipLot, Long> {
 }
