+++ BE/amr_control_system/src/main/java/com/sfaas/amr_control_system/service/WorkHistoryIdentifierHelper.java
@@ -7,11 +7,11 @@ final class WorkHistoryIdentifierHelper {
     private WorkHistoryIdentifierHelper() {
     }
 
-    static String formatWorkHistoryId(Integer taskId) {
+    static String formatWorkHistoryId(Long taskId) {
         return String.format("%s%03d", ID_PREFIX, taskId);
     }
 
-    static Integer parseWorkHistoryId(String workHistoryId) {
+    static Long parseWorkHistoryId(String workHistoryId) {
         if (workHistoryId == null || workHistoryId.isBlank()) {
             throw new IllegalArgumentException("workHistoryId is required.");
         }
@@ -20,7 +20,7 @@ final class WorkHistoryIdentifierHelper {
             throw new IllegalArgumentException("Invalid workHistoryId format. Expected example: wh-001");
         }
         try {
-            return Integer.parseInt(normalized.substring(ID_PREFIX.length()));
+            return Long.parseLong(normalized.substring(ID_PREFIX.length()));
         } catch (NumberFormatException exception) {
             throw new IllegalArgumentException("Invalid workHistoryId format. Expected example: wh-001");
         }
